commit
By using bcrypt, you ensure that the password is securely hashed before storing it in the database during registration. During login, you compare the hashed password from the database with the hashed version of the user's provided password. If they match, the login is successful.

Please note that bcrypt automatically handles the salt generation, which adds an extra layer of security to the password hashing process. It is considered a best practice for securely storing user passwords.